+++ TheOracle2/AskTheOracle/AskOption.cs
@@ -1,6 +1,16 @@
 using System.ComponentModel.DataAnnotations;
-
 namespace TheOracle2;
+public enum AskOption
+{
+  [Display(Name = "Sure thing")]
+  SureThing = 90,
+  Likely = 75,
+  [Display(Name = "Fifty-fifty")]
+  FiftyFifty = 50,
+  Unlikely = 25,
+  [Display(Name = "Small chance")]
+  SmallChance = 10
+}
 
 public class AskDefs
 {
@@ -19,17 +29,5 @@ public class AskDefs
     };
     ByString = new();
   }
-}
-
-public enum AskOption
-{
-  [Display(Name = "Sure thing")]
-  SureThing = 90,
-  Likely = 75,
-  [Display(Name = "Fifty-fifty")]
-  FiftyFifty = 50,
-  Unlikely = 25,
-  [Display(Name = "Small chance")]
-  SmallChance = 10
 }
