+++ TheOracle2/Clock/ClockSize.cs
