+++ TheOracle2/Commands/ClockCommand.cs
@@ -1,16 +1,16 @@
 using Discord.Interactions;
-using Discord.Net;
 using Discord.WebSocket;
 using TheOracle2.GameObjects;
 namespace TheOracle2;
 public class ClockCommand : InteractionModuleBase
 {
-  [SlashCommand("clock", "Creates a clock or scene challenge (see p. 230).")]
+
+  [SlashCommand("clock", "Set a tension clock, campaign clock, or scene challenge (p. 230)")]
   public async Task BuildClock(
     [Summary(description: "The type of clock.")]
     [Choice("Tension clock", "tension-clock"),
     Choice("Campaign clock", "campaign-clock"),
-    // Choice("Scene challenge", "scene-challenge")
+    Choice("Scene challenge", "scene-challenge")
     ] string type,
     [Summary(description: "The number of segments for the clock")]
     [Choice("4 segments", 4),
@@ -44,45 +44,39 @@ public class ClockCommand : InteractionModuleBase
         break;
     }
   }
+
   [ComponentInteraction("reset-clock")]
   public async Task ResetClock()
   {
     SocketMessageComponent interaction = Context.Interaction as SocketMessageComponent;
-    Embed embed = interaction.Message.Embeds.FirstOrDefault();
-    Clock clock = Clock.FromEmbed(embed);
-    clock.Reset();
+    IClock clock = IClock.FromEmbed(interaction.Message.Embeds.FirstOrDefault());
     await interaction.UpdateAsync(msg =>
   {
+    clock.Filled = 0;
     msg.Components = clock.MakeComponents().Build();
     msg.Embed = clock.ToEmbed().Build();
   });
-
   }
   [ComponentInteraction("advance-clock")]
   public async Task AdvanceClock()
   {
     SocketMessageComponent interaction = Context.Interaction as SocketMessageComponent;
-    // SocketMessageComponentData data = interaction.Data;
-    Embed embed = interaction.Message.Embeds.FirstOrDefault();
-    Clock clock = Clock.FromEmbed(embed);
-    clock.Advance();
+    IClock clock = IClock.FromEmbed(interaction.Message.Embeds.FirstOrDefault());
     await interaction.UpdateAsync(msg =>
     {
+      clock.Filled++;
       msg.Components = clock.MakeComponents().Build();
       msg.Embed = clock.ToEmbed().Build();
     });
+    await interaction.FollowupAsync(embed: clock.AlertEmbed().Build());
   }
 
-
   [ComponentInteraction("advance-clock-menu")]
-  public async Task AdvanceClockMenu(string customId, string[] values)
+  public async Task AdvanceClockMenu(string[] values)
   {
-
     SocketMessageComponent interaction = Context.Interaction as SocketMessageComponent;
+    IClock clock = IClock.FromEmbed(interaction.Message.Embeds.FirstOrDefault());
     string optionValue = values.FirstOrDefault();
-    Embed embed = interaction.Message.Embeds.FirstOrDefault();
-    Clock clock = Clock.FromEmbed(embed);
-
     if (int.TryParse(optionValue.Replace("advance-", ""), out int odds))
     {
       OracleAnswer answer = new(odds, $"Does the clock *{clock.Text}* advance?");
@@ -93,76 +87,42 @@ public class ClockCommand : InteractionModuleBase
         {
           answerEmbed.WithFooter("You rolled a match! Two segments have been marked on the clock. Envision how this situation or project gains dramatic support or inertia.");
         }
-        clock.Advance(answer.IsMatch ? 2 : 1);
         await interaction.UpdateAsync(msg =>
         {
+          clock.Filled += answer.IsMatch ? 2 : 1;
           msg.Embed = clock.ToEmbed().Build();
           msg.Components = clock.MakeComponents().Build();
         });
-        answerEmbed = answerEmbed.WithThumbnailUrl(clock.GetImage());
+        answerEmbed = answerEmbed.WithThumbnailUrl(IClock.Images[clock.Segments].ElementAt(clock.Filled));
         await interaction.FollowupAsync(embed: answerEmbed.Build());
         return;
       }
-
       if (answer.IsMatch)
       {
         answerEmbed.WithFooter("You rolled a match! Envision a surprising turn of events which pits new factors or forces against the clock.");
       }
-      answerEmbed = answerEmbed.WithThumbnailUrl(clock.GetImage());
+      answerEmbed = answerEmbed.WithThumbnailUrl(IClock.Images[clock.Segments].ElementAt(clock.Filled));
       await interaction.RespondAsync(
         embed: answerEmbed.Build());
       return;
     }
-
     switch (optionValue)
     {
       case "reset":
-        clock.Reset();
+        clock.Filled = 0;
         break;
       case "advance":
-        clock.Advance();
+        clock.Filled++;
         break;
       default:
         throw new ArgumentOutOfRangeException(nameof(optionValue), "Could not parse integer or valid string from select menu option value.");
-        break;
     }
-
     await interaction.UpdateAsync(msg =>
-  {
-    msg.Components = clock.MakeComponents().Build();
-    msg.Embed = clock.ToEmbed().Build();
-  });
+    {
+      msg.Components = clock.MakeComponents().Build();
+      msg.Embed = clock.ToEmbed().Build();
+    });
+    await interaction.FollowupAsync(embed: clock.AlertEmbed().Build());
   }
-
-
-
-  // [ComponentInteraction("advance-clock-*")]
-  // public async Task AdvanceCampaignClock(string oddsString)
-  // {
-  //   int odds = int.Parse(oddsString);
-  //   SocketMessageComponent interaction = Context.Interaction as SocketMessageComponent;
-  //   Console.WriteLine($"interaction.Data.Values.FirstOrDefault(): {interaction.Data.Values.FirstOrDefault()}");
-  //   Embed originalEmbed = interaction.Message.Embeds.FirstOrDefault();
-  //   string originalEmbedTitle = originalEmbed.Title;
-  //   string clockTitle = originalEmbed.Title.Length > 0 ? originalEmbed.Title : originalEmbed.Author.ToString();
-  //   OracleAnswer answer = new(odds, $"Does *{clockTitle}* advance?");
-  //   switch (answer.IsYes)
-  //   {
-  //     case true:
-  //       CampaignClock clock = new(interaction.Message.Embeds.FirstOrDefault());
-  //       int wedgesFilled = answer.IsMatch ? 2 : 1;
-  //       clock.Advance(wedgesFilled);
-  //       await interaction.UpdateAsync(msg =>
-  //         {
-  //           msg.Components = clock.MakeComponents().Build();
-  //           msg.Embed = clock.ToEmbed().Build();
-  //         });
-  //       await interaction.FollowupAsync(embed: answer.ToEmbed().Build());
-  //       break;
-  //     case false:
-  //       await interaction.RespondAsync(embed: answer.ToEmbed().Build());
-  //       break;
-  //   }
-  // }
 }
