--- TheOracle2/Images/Clocks.cs
@@ -1,58 +1,0 @@
-using Discord;
-
-namespace TheOracle2.Images;
-
-public class ClockImages : Dictionary<int, string[]>
-{
-
-  public ClockImages()
-  // imgur gallery: https://imgur.com/a/uf9oBgT
-  {
-    Add(
-      4,
-      new string[]{
-        "https://i.imgur.com/Ahbi1DV.png",
-        "https://i.imgur.com/7RTHuPB.png",
-        "https://i.imgur.com/a0hlbCn.png",
-        "https://i.imgur.com/A3e5aHC.png",
-        "https://i.imgur.com/8DzJNyy.png",
-      }
-    );
-    Add(6,
-      new string[]{
-      "https://i.imgur.com/rvDLRZO.png",
-      "https://i.imgur.com/EZbNNRC.png",
-      "https://i.imgur.com/RlaRzgz.png",
-      "https://i.imgur.com/OJ2WVSR.png",
-      "https://i.imgur.com/YgLlojT.png",
-      "https://i.imgur.com/ZpyIpTC.png",
-      "https://i.imgur.com/KjYg7aC.png"
-      }
-    );
-    Add(8,
-    new string[]{
-      "https://i.imgur.com/Qi0pkYD.png",
-      "https://i.imgur.com/9wwW3Sh.png",
-      "https://i.imgur.com/sM27Mbf.png",
-      "https://i.imgur.com/GsKElZv.png",
-      "https://i.imgur.com/QvYHujk.png",
-      "https://i.imgur.com/nNEFyRr.png",
-      "https://i.imgur.com/a9Qvkin.png",
-      "https://i.imgur.com/0dclJ9Q.png",
-      "https://i.imgur.com/5Z0bQ9K.png",
-    });
-    Add(10, new string[]{
-      "https://i.imgur.com/u9erdAx.png",
-      "https://i.imgur.com/0zfi1PJ.png",
-      "https://i.imgur.com/ayanbMK.png",
-      "https://i.imgur.com/OtfwmEf.png",
-      "https://i.imgur.com/uUWoyZV.png",
-      "https://i.imgur.com/eKhfGoj.png",
-      "https://i.imgur.com/cwzEkCD.png",
-      "https://i.imgur.com/GNJPzru.png",
-      "https://i.imgur.com/cMQNAZV.png",
-      "https://i.imgur.com/EaOxmdt.png",
-      "https://i.imgur.com/NRhTaBR.png",
-    });
-  }
-}
